+++ src/com/org/gnos/db/model/OreMiningCost.java
@@ -6,7 +6,7 @@ public class OreMiningCost extends FixedOpexCost{
 	public OreMiningCost(){
 		super();
 		this.id = -1;
-		this.costData = new LinkedHashMap<Integer, Integer>();
+		this.costData = new LinkedHashMap<Integer, Float>();
 
 	}
